+++ lib/wave_generator.dart
@@ -7,22 +7,33 @@ import 'package:wave_generator/src/data_chunk8.dart';
 import 'package:wave_generator/src/format_chunk.dart';
 import 'package:wave_generator/src/wave_header.dart';
 
+/// Bit-depth per sample.
 enum BitDepth {
   Depth8bit,
 //  Depth16bit,
 //  Depth32bit
 }
 
+/// Waveform for a tone.
 enum Waveform {
   Sine,
   Square,
   Triangle,
 }
 
+/// Represents a single tone.
 class Note {
+
+  /// Frequency in Hz.
   final double frequency;
+
+  /// Duration in milliseconds.
   final int msDuration;
+
+  /// Waveform of the tone.
   final Waveform waveform;
+
+  /// Volume in the range 0.0 - 1.0.
   final double volume;
 
   Note(this.frequency, this.msDuration, this.waveform, this.volume){
@@ -39,13 +50,17 @@ class Note {
     return Note(440.0, duration, Waveform.Sine, volume);
   }
 
-// Etc
+// Etc, do more
 // http://pages.mtu.edu/~suits/notefreqs.html
 }
 
+/// Generates simple waveforms as uncompressed PCM audio data.
 class WaveGenerator
 {
+  /// Samples generated per second.
   final int sampleRate;
+
+  /// Bit-depth of each audio sample.
   final BitDepth bitDepth;
 
   factory WaveGenerator.simple() {
@@ -54,6 +69,7 @@ class WaveGenerator
 
   WaveGenerator(this.sampleRate, this.bitDepth);
 
+  /// Generate a byte stream equivalent to a wav file of the Note argument
   Stream<int> generate(Note note) async* {
 
     var formatHeader = FormatChunk(1, sampleRate, bitDepth);
@@ -75,11 +91,11 @@ class WaveGenerator
     }
   }
 
+  /// Generate a byte stream equivalent to a wav file of the Note list argument, played sequentially
   Stream<int> generateSequence(List<Note> notes) async* {
+
     var formatHeader = FormatChunk(1, sampleRate, bitDepth);
-
     var dataChunk = _getDataChunk(formatHeader, notes );
-
     var fileHeader = WaveHeader(formatHeader, dataChunk);
 
     await for (int data in fileHeader.bytes()) {